# [E2E] Activate block theme via the rest API instead via CLI