--- a/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
+++ b/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
@@ -1,24 +1,25 @@
 /**
  * External dependencies
  */
-import {
-	BLOCK_THEME_NAME,
-	BLOCK_THEME_SLUG,
-	cli,
-} from '@woocommerce/e2e-utils';
-import { test as setup, expect } from '@woocommerce/e2e-playwright-utils';
+import { BLOCK_THEME_SLUG, cli } from '@woocommerce/e2e-utils';
+import { test as setup } from '@woocommerce/e2e-playwright-utils';
 
-setup( 'Sets up the block theme', async ( { admin } ) => {
-	await cli(
-		`npm run wp-env run tests-cli -- wp theme install ${ BLOCK_THEME_SLUG } --activate`
-	);
-	await admin.page.goto( '/wp-admin/themes.php' );
-	await expect(
-		admin.page.getByText( `Active: ${ BLOCK_THEME_NAME }` )
-	).toBeVisible();
-	// Enable permalinks.
-	await cli(
-		`npm run wp-env run tests-cli -- wp rewrite structure /%postname%/ --hard`
-	);
-	await cli( `npm run wp-env run tests-cli -- wp rewrite flush --hard` );
+type ThemeItem = {
+	status: string;
+	stylesheet: string;
+};
+
+setup( 'Sets up the block theme', async ( { requestUtils } ) => {
+	const themes = await requestUtils.rest< ThemeItem[] >( {
+		path: '/wp/v2/themes',
+	} );
+	const currentTheme = themes.find( ( { status } ) => status === 'active' );
+
+	if ( ! currentTheme || currentTheme.stylesheet !== BLOCK_THEME_SLUG ) {
+		await requestUtils.activateTheme( BLOCK_THEME_SLUG );
+		await cli(
+			`npm run wp-env run tests-cli -- wp rewrite structure /%postname%/ --hard`
+		);
+		await cli( `npm run wp-env run tests-cli -- wp rewrite flush --hard` );
+	}
 } );