--- a/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
+++ b/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
@@ -2,21 +2,23 @@
  * External dependencies
  */
 import { BLOCK_THEME_SLUG, cli } from '@woocommerce/e2e-utils';
-import { test as setup } from '@woocommerce/e2e-playwright-utils';
+import { test as setup, expect } from '@woocommerce/e2e-playwright-utils';
 
-type ThemeItem = {
-	status: string;
-	stylesheet: string;
-};
+setup( 'Sets up the block theme', async () => {
+	const isActiveOutput = await cli(
+		`npm run wp-env run tests-cli -- wp theme is-active ${ BLOCK_THEME_SLUG }`
+	);
 
-setup( 'Sets up the block theme', async ( { requestUtils } ) => {
-	const themes = await requestUtils.rest< ThemeItem[] >( {
-		path: '/wp/v2/themes',
-	} );
-	const currentTheme = themes.find( ( { status } ) => status === 'active' );
+	if ( isActiveOutput.stderr.includes( 'Command failed' ) ) {
+		const installOutput = await cli(
+			`npm run wp-env run tests-cli -- wp theme install ${ BLOCK_THEME_SLUG } --activate`
+		);
 
-	if ( ! currentTheme || currentTheme.stylesheet !== BLOCK_THEME_SLUG ) {
-		await requestUtils.activateTheme( BLOCK_THEME_SLUG );
+		expect(
+			installOutput.stdout,
+			`Could not install and/or activate ${ BLOCK_THEME_SLUG }`
+		).toContain( 'Success' );
+
 		await cli(
 			`npm run wp-env run tests-cli -- wp rewrite structure /%postname%/ --hard`
 		);