--- a/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
+++ b/plugins/woocommerce-blocks/tests/e2e/block-theme.setup.ts
@@ -5,18 +5,20 @@ import { BLOCK_THEME_SLUG, cli } from '@woocommerce/e2e-utils';
 import { test as setup, expect } from '@woocommerce/e2e-playwright-utils';
 
 setup( 'Sets up the block theme', async () => {
-	const { stdout } = await cli(
+	let cliOutput = await cli(
 		`npm run wp-env run tests-cli -- wp theme install ${ BLOCK_THEME_SLUG } --activate`
 	);
 
 	expect(
-		stdout,
+		cliOutput.stdout,
 		`Could not install and/or activate ${ BLOCK_THEME_SLUG }`
 	).toContain( 'Success' );
 
-	// Enable permalinks.
-	await cli(
+	// Enable permalinks and perform a hard flush.
+	cliOutput = await cli(
 		`npm run wp-env run tests-cli -- wp rewrite structure /%postname%/ --hard`
 	);
-	await cli( `npm run wp-env run tests-cli -- wp rewrite flush --hard` );
+
+	expect( cliOutput.stdout ).toContain( 'Success: Rewrite structure set' );
+	expect( cliOutput.stdout ).toContain( 'Success: Rewrite rules flushed' );
 } );