--- a/plugins/woocommerce-blocks/tests/e2e/utils/cli.ts
+++ b/plugins/woocommerce-blocks/tests/e2e/utils/cli.ts
@@ -1,10 +1,17 @@
 /**
  * External dependencies
  */
-import { exec } from 'child_process';
+import { ExecException, exec } from 'child_process';
 
-export function cli( cmd, args = [] ) {
-	return new Promise( ( resolve ) => {
+type CliOutput = {
+	code: number;
+	error: ExecException | null;
+	stdout: string;
+	stderr: string;
+};
+
+export function cli( cmd: string, args = [] ) {
+	return new Promise< CliOutput >( ( resolve ) => {
 		exec( `${ cmd } ${ args.join( ' ' ) }`, ( error, stdout, stderr ) => {
 			resolve( {
 				code: error && error.code ? error.code : 0,