### Changes proposed in this Pull Request:

In the `block-theme.setup.ts`, activate the theme via the rest API instead via CLI. This approach should provide faster and more stable behavior as it doesn't require any manual actions against the browser, like setting the theme via the themes page. 

In terms of the speed improvement, here are some rough numbers:
- For initial call it improves the speed from ~8.5s to ~5.3s (~38%).
- For subsequent calls, it goes down from ~8.5s to ~230ms (~97%) by recognizing the theme is already set.

⚠️ The absolute measurements may vary depending on the machine!


### How to test the changes in this Pull Request:

1. Manually, change the theme on the localhost:8889 instance to, e.g., Storefront
2. Run some test that uses the block_theme.setup.ts, e.g. `price-filter.block_theme.side_effects.spec.ts`
3. Ensure the theme has been set correctly for the test.
4. Run the test again and ensure the theme is still correct :)